Conflicts:
	src/openMVG/sfm/pipelines/sequential/sequential_SfM.cpp
	src/software/SfM/SfMPlyHelper.hpp

+++ src/openMVG/image/image_io_test.cpp
@@ -226,7 +226,7 @@ TEST(ImageHeader, AllFormats) {
 
   const std::vector<std::string> ext_Type = {"jpg", "png", "tif", "png", "pgm"};
   const int image_border_size = 10;
-  for (int i=0; i < ext_Type.size(); ++i)
+  for (size_t i=0; i < ext_Type.size(); ++i)
   {
     std::ostringstream os;
     os << "img" << "." << ext_Type[i];
